+++ .github/workflows/check.yml
@@ -24,8 +24,3 @@ jobs:
       - run: pnpm oxfmt --check
       - run: pnpm build
       - run: pnpm test
-        env:
-          CDP_API_KEY_ID: ${{ vars.CDP_API_KEY_ID }}
-          CDP_API_KEY_SECRET: ${{ secrets.CDP_API_KEY_SECRET }}
-          EVM_SEED_PHRASE: ${{ secrets.EVM_SEED_PHRASE }}
-          PAY_TO_EVM: ${{ vars.PAY_TO_EVM }}
