+++ @crosshatch/facilitator/vitest.config.ts
@@ -1,4 +1,4 @@
-import Path from "node:path"
+import Process from "node:process"
 
 import { loadEnv } from "vite"
 import { mergeConfig, type ViteUserConfig } from "vitest/config"
@@ -8,6 +8,6 @@ import config from "../../konfik/vitest.ts"
 export default mergeConfig(config, {
   test: {
     name: "@crosshatch/facilitator",
-    env: loadEnv("test", Path.resolve(import.meta.dirname, "../.."), ""),
+    env: loadEnv("test", Process.cwd(), ""),
   },
 } satisfies ViteUserConfig)
