--- @crosshatch/merchant/Settlement.ts
@@ -1,3 +1,0 @@
-import { Context } from "effect"
-
-export class Settlement extends Context.Service<Settlement, {}>()("@crosshatch/merchant/Settlement") {}
