+++ crosshatch/Trace.ts
@@ -1,36 +1,14 @@
-import { Bridge } from "crosshatch"
-import { Context, Option, Effect, flow, String } from "effect"
+import { Option, Schema as S, Context, Effect, flow } from "effect"
 
-export class Trace extends Context.Service<
-  Trace,
-  {
-    readonly traceId: string
-    readonly name: string
-    readonly description: string
-  }
->()("@crosshatch/merchant/Trace") {}
+export const TraceConfig = S.Struct({
+  traceId: S.String,
+  name: S.String,
+  description: S.String,
+})
 
-export const traced =
-  (name: string) =>
-  (template: TemplateStringsArray, ...substitutions: ReadonlyArray<unknown>) =>
-    Effect.fnUntraced(function* <A, E, R>(effect: Effect.Effect<A, E, R>) {
-      const { createTrace } = yield* Bridge
-      const traceId = yield* Effect.currentSpan.pipe(
-        Effect.map(({ traceId }) => traceId),
-        Effect.catchTags({
-          NoSuchElementError: Effect.die,
-        }),
-      )
-      const trace = {
-        traceId,
-        name,
-        description: String.stripMargin(globalThis.String.raw(template, ...substitutions)),
-      }
-      yield* createTrace(trace)
-      return yield* Effect.provideService(effect, Trace, trace)
-    }, Effect.withSpan(name))
+export class Trace extends Context.Service<Trace, typeof TraceConfig.Type>()("crosshatch/Trace") {}
 
-export const CurrentTraceId = Effect.serviceOption(Trace).pipe(
+export const TraceId = Effect.serviceOption(Trace).pipe(
   Effect.map(
     flow(
       Option.map(({ traceId }) => traceId),
