+++ crosshatch/index.ts
@@ -5,13 +5,14 @@ export * from "./LinkChallengeId.ts"
 export * from "./atoms.ts"
 export * from "./http.ts"
 export * from "./makeFetch.ts"
-export * from "./fns.ts"
+export * from "./wallet.ts"
 export { CrosshatchLive } from "./runtime.ts"
-export * from "./TraceConfig.ts"
-export * from "./Trace.ts"
-export * from "./Invoice.ts"
 export * as Amount from "./Amount.ts"
+export * from "./traced.ts"
 export * as Asset from "./Asset.ts"
-export * from "./Bridge.ts"
+export * as Payer from "./Payer.ts"
 export * from "./errors.ts"
 export * as Prerequisite from "./Prerequisite.ts"
+export * as Trace from "./Trace.ts"
+export * from "./createPayload.ts"
+export * as Invoice from "./Invoice.ts"
