+++ crosshatch/wallet.ts
@@ -1,5 +1,5 @@
 import * as Facade from "./Facade/Facade.ts"
 
-export const propose = Facade.FacadeClient.fn("Propose")
-
 export const createTrace = Facade.FacadeClient.fn("CreateTrace")
+
+export const propose = Facade.FacadeClient.fn("Propose")
